+++ far/panel.cpp
@@ -382,7 +382,7 @@ static void ConfigureChangeDriveMode()
 }
 
 
-intptr_t WINAPI ChDiskDlgProc(HANDLE hDlg,int Msg,int Param1,void* Param2)
+intptr_t WINAPI ChDiskDlgProc(HANDLE hDlg,intptr_t Msg,intptr_t Param1,void* Param2)
 {
 	switch (Msg)
 	{
