+++ far/viewer.cpp
@@ -134,7 +134,7 @@ Viewer::Viewer(bool bQuickView, UINT aCodePage):
 	vgetc_cb = vgetc_ib = 0;
 	vgetc_composite = L'\0';
 
-	vread_buffer_size = Max(MAX_VIEWLINEB, 8192);
+	vread_buffer_size = Max(MAX_VIEWLINEB, (intptr_t)8192);
 	vread_buffer = new char[vread_buffer_size];
 
 	lcache_first = lcache_last = -1;
@@ -144,14 +144,14 @@ Viewer::Viewer(bool bQuickView, UINT aCodePage):
 	lcache_ready = false;
 	lcache_wrap = lcache_wwrap = lcache_width = -1;
 
-	int cached_buffer_size = 64*Max(Opt.ViOpt.MaxLineSize*2, 1024);
+	int cached_buffer_size = 64*Max(Opt.ViOpt.MaxLineSize*2, (intptr_t)1024);
 	max_backward_size = ViewerOptions::eMaxLineSize*3;
 	if ( max_backward_size > cached_buffer_size/2 )
 		max_backward_size = cached_buffer_size / 2;
 	llengths_size = max_backward_size / 40;
 	llengths = new int[llengths_size];
 
-	Search_buffer_size = 3 * Max(MAX_VIEWLINEB, 8000);
+	Search_buffer_size = 3 * Max(MAX_VIEWLINEB, (intptr_t)8000);
 	Search_buffer = new wchar_t[Search_buffer_size];
 
 	ClearStruct(vString);
@@ -2468,7 +2468,7 @@ struct MyDialogData
 	bool      recursive;
 };
 
-intptr_t WINAPI ViewerSearchDlgProc(HANDLE hDlg,int Msg,int Param1,void* Param2)
+intptr_t WINAPI ViewerSearchDlgProc(HANDLE hDlg,intptr_t Msg,intptr_t Param1,void* Param2)
 {
 	switch (Msg)
 	{
